% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjSameName.R
\name{adjSameName}
\alias{adjSameName}
\title{adjSameName}
\usage{
adjSameName(adjL)
}
\arguments{
\item{adjL}{list of adjacency matrices}
}
\value{
a list of adjacency matrices with the SAME nodes in rows and columns
}
\description{
it construct a list of adjacency matrices with the SAME nodes
in rows and columns
}
\examples{
data("tryL_data")
adjSameName(tryL_data)

}
