% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityWeight.R
\name{densityNet}
\alias{densityNet}
\title{densityNet}
\usage{
densityNet(graphsL)
}
\arguments{
\item{graphsL}{the list of weighted graphs in igraph format.}
}
\value{
- the quantile of the mean density distribution
- the quantile of the mean density distribution without the zeros
- plot density distribution without the zeros
}
\description{
The function creates a density plot of the different graphs mean
weights. It can be used to search the final Threshold for the Consensus
Network starting from similar networks.
}
\examples{
gI <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gI)$weight <- runif(igraph::ecount(gI),0.5,1)
gII <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gII)$weight <- runif(igraph::ecount(gII),0.5,1)

graphsList <- list(gI,gII)

densityNet(graphsL=graphsList)

}
