% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityWeight.R
\name{DensityWeight}
\alias{DensityWeight}
\title{DensityWeight}
\usage{
DensityWeight(graphsL)
}
\arguments{
\item{graphsL}{the list of weighted graphs in igraph format.}
}
\value{
- the quantile of the mean density distribution
- the quantile of the mean density distribution without the zeros
- plot density distribution without the zeros
}
\description{
density plot of the mean of the weights of different graphs. It
can be used to search the final Threshold for the Consensus Network
construction from the similar networks (output "similarGraphs" of the
ConsensusAlgorithm function, default Threshold is 0.5).
}
\examples{
gI <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gI)$weight <- runif(igraph::ecount(gI),0.5,1)
gII <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gII)$weight <- runif(igraph::ecount(gII),0.5,1)

graphsList <- list(gI,gII)

DensityWeight(graphsL=graphsList)

}
