% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardWeightedMatrix.R
\name{JaccardWeightedMatrixDistance}
\alias{JaccardWeightedMatrixDistance}
\title{JaccardWeightedMatrixDistance}
\usage{
JaccardWeightedMatrixDistance(graphsL)
}
\arguments{
\item{graphsL}{list of graphs as igraph objects with the SAME nodes.}
}
\value{
Weighted Jaccard Distance Matrix.
}
\description{
Jaccard Weighted Matrix Distance between all the graphs in
pairs.
}
\examples{
gI <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gI)$weight <- runif(igraph::ecount(gI),0.5,1)
gII <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gII)$weight <- runif(igraph::ecount(gII),0.5,1)


graphsList <- list(gI,gII)
JaccardWeightedMatrixDistance(graphsL=graphsList)

}
