% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardWeightedMean.R
\name{JWmean}
\alias{JWmean}
\title{JWmean}
\usage{
JWmean(graphsL)
}
\arguments{
\item{graphsL}{list of different graphs in igraph format with SAME nodes.}
}
\value{
a number.
}
\description{
Jaccard Weighted Mean Distance to understand how different are
the layers.
}
\examples{
gI <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gI)$weight <- runif(igraph::ecount(gI),0.5,1)
gII <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gII)$weight <- runif(igraph::ecount(gII),0.5,1)


graphsList <- list(gI,gII)
JWmean(graphsL=graphsList)
}
