% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkMeasures.R
\name{measuresNet}
\alias{measuresNet}
\title{measuresNet}
\usage{
measuresNet(graphsL, NodesMeasures = TRUE)
}
\arguments{
\item{graphsL}{a list of graphs as igraphs objects.}

\item{NodesMeasures}{logical, if falso it computes only graph measures, if
true it computes also nodes measures (default TRUE).}
}
\value{
list of measure for each layer.
}
\description{
graphs and nodes measures to analyse all the layers in one shot.
}
\examples{
gI <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gI)$weight <- runif(igraph::ecount(gI),0.5,1)
gII <- igraph::sample_pa(n=10,directed=FALSE)
igraph::E(gII)$weight <- runif(igraph::ecount(gII),0.5,1)
graphsList <- list(gI,gII)
measuresNet(graphsL=graphsList, NodesMeasures=TRUE)

}
