% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotL}
\alias{plotL}
\title{plotL}
\usage{
plotL(graphL, ...)
}
\arguments{
\item{graphL}{List of graphs}

\item{...}{other parameter}
}
\value{
plot of graphs
}
\description{
plot the graphs
}
\examples{
data("graphL_data")
plotL(graphL_data)
}
