% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardWeightedMatrix.R
\name{JWmatrix}
\alias{JWmatrix}
\title{JWmatrix}
\usage{
JWmatrix(graphL)
}
\arguments{
\item{graphL}{list of graphs as igraph objects with the SAME nodes.}
}
\value{
Weighted Jaccard Distance Matrix.
}
\description{
Jaccard Weighted Matrix Distance between all the graphs in
pairs.
}
\examples{
data("graphL_data")
JWmatrix(graphL_data)
}
