% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardWeightedMean.R
\name{JWmean}
\alias{JWmean}
\title{JWmean}
\usage{
JWmean(graphL)
}
\arguments{
\item{graphL}{list of different graphs in igraph format with SAME nodes.}
}
\value{
a number.
}
\description{
Jaccard Weighted Mean Distance to understand how different are
the layers.
}
\examples{
data("graphL_data")
JWmean(graphL_data)
}
