% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecificFunction.R
\name{specificNet}
\alias{specificNet}
\title{specificNet}
\usage{
specificNet(adjL, graph.consensus)
}
\arguments{
\item{adjL}{list of weighted adjacency matrix with the same names in rows
and columns for all the matrices.}

\item{graph.consensus}{graphConsensus output of the
\code{\link{consensusNet}} function.}
}
\value{
Case Specific Networks one for each layer and percentage of
specificity.
}
\description{
The function creates Case Specific Networks one for each layer
to give information of the peculiar layer not present in the Consensus.
}
\examples{
data("adjL_data")
myConsensus <- consensusNet(adjL_data)
specificNet(adjL_data, myConsensus$graphConsensus)
}
